commit at 2020-12-28 14:10

+++ src/common/network/request.js
@@ -38,6 +38,7 @@ const codeMessage = {
 }
 // 是否直接在请求成功里去除loding
 let isLoadingFlag = true
+let returnFlag = true
 // 解决快速点击或并发请求出现的多个请求的问题
 const pending = [] // 声明一个数组用于存储每个ajax请求的取消函数和ajax标识
 // eslint-disable-next-line no-unused-vars
@@ -94,9 +95,12 @@ instance.interceptors.response.use(response => {
     return Promise.reject(response.data.errorMsg)
   } else if (response.data.errorCode === 1003) {
     store.commit('SETSPINNING', false)
-    Message.warning('用户身份信息过期，请重新登录')
-    sessionStorage.clear()
-    router.push('./')
+    if (returnFlag) {
+      returnFlag = false
+      Message.warning('用户身份信息过期，请重新登录')
+      sessionStorage.clear()
+      router.push('./')
+    }
   } else {
     return response.data
   }
